# services/ai/openapi.json
{
  "components": {
    "schemas": {
      "FxRate": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "base": {
            "title": "Base",
            "type": "string"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "rate": {
            "description": "Decimal as string to avoid float drift in transport.",
            "title": "Rate",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "base",
          "quote",
          "rate",
          "as_of",
          "source"
        ],
        "title": "FxRate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "checks": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Checks",
            "type": "object"
          },
          "service": {
            "default": "ai-service",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "Instrument": {
        "properties": {
          "asset_class": {
            "default": "unknown",
            "enum": [
              "equity",
              "etf",
              "crypto",
              "fx",
              "index",
              "unknown"
            ],
            "title": "Asset Class",
            "type": "string"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source": {
            "default": "unknown",
            "title": "Source",
            "type": "string"
          },
          "symbol": {
            "description": "Canonical provider symbol, e.g. 'AAPL', 'BTC-USD'.",
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "Instrument",
        "type": "object"
      },
      "InstrumentResolution": {
        "description": "Result of resolving a user-supplied string to a tradable instrument.\n\n`candidates` is non-empty and ordered by confidence whenever `resolved` is\nNone, so the UI can ask the user to disambiguate instead of guessing.",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Instrument"
            },
            "title": "Candidates",
            "type": "array"
          },
          "confidence": {
            "default": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "resolved": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instrument"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "query"
        ],
        "title": "InstrumentResolution",
        "type": "object"
      },
      "Quote": {
        "description": "A single price observation. `price_minor` is integer minor units.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "day_change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Change Pct"
          },
          "delay_seconds": {
            "default": 0,
            "description": "Provider-declared quote delay. Displayed in the UI; 0 means real time.",
            "title": "Delay Seconds",
            "type": "integer"
          },
          "previous_close_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close Minor"
          },
          "price_minor": {
            "title": "Price Minor",
            "type": "integer"
          },
          "source": {
            "description": "Provider that produced this quote, e.g. 'fixture', 'yfinance'.",
            "title": "Source",
            "type": "string"
          },
          "stale": {
            "default": false,
            "description": "True when served from cache after every live provider failed.",
            "title": "Stale",
            "type": "boolean"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "price_minor",
          "currency",
          "as_of",
          "source"
        ],
        "title": "Quote",
        "type": "object"
      },
      "QuoteRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "symbols"
        ],
        "title": "QuoteRequest",
        "type": "object"
      },
      "QuoteResponse": {
        "properties": {
          "missing": {
            "description": "Symbols no provider in the chain could price.",
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/Quote"
            },
            "title": "Quotes",
            "type": "array"
          }
        },
        "required": [
          "quotes"
        ],
        "title": "QuoteResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Market data, analysis pipeline and RAG engine for the portfolio copilot.",
    "title": "Traders AI Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz",
        "tags": [
          "health"
        ]
      }
    },
    "/market/fx": {
      "get": {
        "operationId": "get_fx_rate_market_fx_get",
        "parameters": [
          {
            "in": "query",
            "name": "base",
            "required": true,
            "schema": {
              "maxLength": 8,
              "minLength": 3,
              "title": "Base",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quote",
            "required": true,
            "schema": {
              "maxLength": 8,
              "minLength": 3,
              "title": "Quote",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FxRate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Fx Rate",
        "tags": [
          "market"
        ]
      }
    },
    "/market/instruments/resolve": {
      "get": {
        "operationId": "resolve_instrument_market_instruments_resolve_get",
        "parameters": [
          {
            "description": "Symbol or company name",
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "description": "Symbol or company name",
              "maxLength": 64,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentResolution"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Instrument",
        "tags": [
          "market"
        ]
      }
    },
    "/market/quotes": {
      "post": {
        "operationId": "get_quotes_market_quotes_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-internal-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quotes",
        "tags": [
          "market"
        ]
      }
    },
    "/readyz": {
      "get": {
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readyz",
        "tags": [
          "health"
        ]
      }
    }
  }
}